traceo/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.1.0
aiosmtplib==3.0.1
jinja2==3.1.2
requests==2.31.0
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# IMAP email handling
imap-tools==1.5.0
email-parser==2.4.0

# IP/Domain information
geoip2==4.7.0
dnspython==2.4.2
python-whois==0.9.4

# Async support
aiofiles==23.2.1

# Logging & monitoring
loguru==0.7.2
